backend/requirements.txt
deepgram-sdk==3.7.2
fastapi==0.128.0
langchain==0.3.14
langchain-chroma==0.2.2
langchain-community==0.3.14
langchain-core==0.3.63
langchain-google-genai==2.0.9
langchain-huggingface==0.2.0
langchain-text-splitters==0.3.5
protobuf==5.29.5
pydantic==2.12.5
pymongo==4.15.5
PyPDF2==3.0.1
python-dotenv==1.2.1
bcrypt==5.0.0
python-jose==3.5.0
python-multipart==0.0.21
requests==2.32.5
twilio==9.9.1